Refactor importlib metadata backend to subpackage
