+++ src/pip/_internal/metadata/importlib/__init__.py
@@ -1,0 +1,4 @@
+from ._dists import Distribution
+from ._envs import Environment
+
+__all__ = ["Distribution", "Environment"]
